chore(scenes): reorganize scene components into subdirectories
- Move scene components into logical subdirectories:
  - gallery/: DemoScene1, DemoScene2, HallwayScene
  - hallway/: HallwayControlScene
  - serial/: SerialHallwayScene
