+++ src/scenes/gallery/DemoScene2.vue
@@ -6,16 +6,16 @@
 
 <script>
 import { onMounted, onUnmounted, ref } from 'vue'
-import { DemoWorld01 } from '../worlds/DemoWorld01.js'
+import { DemoWorld02 } from '../../worlds/DemoWorld02.js'
 
 export default {
-  name: 'Scene1',
+  name: 'Scene2',
   setup() {
     const canvas = ref(null)
     let world = null
 
     onMounted(() => {
-      world = new DemoWorld01(canvas.value)
+      world = new DemoWorld02(canvas.value)
       world.init()
     })
