+++ src/scenes/hallway/HallwayControlScene.vue
